refactor: move TableNodeContent to separate file
